+++ src/interfaces/IConditionEvaultor.sol
@@ -3,10 +3,10 @@ pragma solidity ^0.8.26;
 
 interface IConditionEvaultor {
     function evaluateCondition(
-        uint8 _platform,
+        uint16 _platform,
         address _platformAddress,
         address _borrower,
-        uint8 _parameter,
+        uint16 _parameter,
         uint256 _conditionValue
     ) external view returns (bool);
 }
